requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.4
django-debug-toolbar==4.4.6
django-environ==0.12.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.2.2
factory_boy==3.3.1
Faker==35.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
Markdown==3.7
packaging==24.2
pluggy==1.5.0
psycopg-binary==3.2.3
psycopg2-binary==2.9.10
pytailwindcss==0.2.0
pytest==8.3.4
pytest-django==4.9.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
